refactor and use es6 availale features
