+++ index.js
@@ -1,15 +1,2 @@
 process.env.CELL_MODE = process.env.CELL_MODE || '_development'
-
-var Cell = require('./lib/cell')
-var instance = new Cell()
-
-instance.start(function (err) {
-  if (err) throw err
-  // # build server cell
-  instance.plasma.emit({type: 'build', branch: 'server.processes.index.plasma'}, function (err) {
-    if (err) throw err
-    instance.plasma.emit({type: 'build', branch: 'server.processes.index.membrane'}, function (err) {
-      if (err) throw err
-    })
-  })
-})
+require('./server')
