--- lib/require.js
@@ -1,4 +1,0 @@
-var path = require('path')
-module.exports = function (modulePath) {
-  return require(path.join(process.cwd(), modulePath))
-}
