+++ package.json
@@ -18,6 +18,7 @@
     "body-parser": "^1.7.0",
     "bunyan": "~1.0.1",
     "cookie-parser": "^1.3.2",
+    "es6-module-loader": "^0.17.9",
     "express": "^4.12.2",
     "glob": "^6.0.1",
     "glob2base": "0.0.12",
@@ -41,6 +42,7 @@
     "organic-express-static": "0.0.2",
     "organic-nucleus": "0.0.3",
     "organic-plasma": "0.0.6",
+    "traceur": "0.0.93",
     "vinyl-named": "^1.1.0"
   },
   "devDependencies": {
