+++ upgrades/client/assetpipeline-webpack/scripts/watchjs.js
@@ -1,7 +1,7 @@
 module.exports = function (angel) {
   angel.on('watchjs', function () {
-    var loadDNA = require('../lib/dna')
-    var runPipeline = require('../lib/gulp-pipeline')
+    var loadDNA = require('../server/load-dna')
+    var runPipeline = require('../server/lib/gulp-pipeline')
     var webpack = require('webpack-stream')
     var sourcemaps = require('gulp-sourcemaps')
     var path = require('path')
