+++ upgrades/release/package.json
@@ -2,6 +2,7 @@
   "devDependencies": {
     "angelscripts-cellcmds": "0.2.2",
     "angelscripts-nginx": "0.0.1",
-    "angelabilities-exec": "0.1.0"
+    "angelabilities-exec": "0.1.0",
+    "string-template": "0.2.1"
   }
 }
